fix(data.procedure.discarded.procedure_paths): was not migrated
